refactor : stub 용 LottoStore 생성
